+++ changelog.md
@@ -5,9 +5,10 @@
 * Statistics
 * Tag_tasks - allows you tag your jobs
 * self.pefiles: introduced to prcessing/signatures modules, you can get PEFILE object by sha256 self.pefiles.get(sha256)
-* https traffic decription
+* https traffic decription WIP
 * Pending page now is much useful and show hashes to easilly spot duplicated
 * Submission of file or resubmission will show all the jobs and detection for that file
+* [Flare capa](https://github.com/fireeye/capa) integrated under static tab(should be enabled in processing.conf under static), Rules can be pulled from community, but we will leave it community driven to sync them. So you can copy them from https://github.com/fireeye/capa-rules and place under `data/flare-capa`
 * More soon ;)
 
 ### 16-11-2020